Move IPA verifier to Halo module
